Parameterize RewriteDataFile's CommitService for RewritePositionDeleteFiles
