+++ .palantir/revapi.yml
@@ -426,6 +426,11 @@ acceptedBreaks:
     - code: "java.field.removedWithConstant"
       old: "field org.apache.iceberg.TableProperties.HMS_TABLE_OWNER"
       justification: "Removing deprecations for 1.3.0"
+    - code: "java.method.parameterTypeChanged"
+      old: "parameter void org.apache.iceberg.actions.RewriteDataFilesCommitManager.CommitService::offer(===org.apache.iceberg.actions.RewriteFileGroup===)"
+      new: "parameter void org.apache.iceberg.actions.BaseCommitService<T>::offer(===T===)\
+        \ @ org.apache.iceberg.actions.RewriteDataFilesCommitManager.CommitService"
+      justification: "Backwards compatible parameterization of argument"
     - code: "java.method.removed"
       old: "method ThisT org.apache.iceberg.BaseScan<ThisT, T extends org.apache.iceberg.ScanTask,\
         \ G extends org.apache.iceberg.ScanTaskGroup<T extends org.apache.iceberg.ScanTask>>::newRefinedScan(org.apache.iceberg.TableOperations,\
